replaced old propertyinfo/macros for core projects
git-svn-id: https://subversion.assembla.com/svn/dsmks/branches/core_refac@661 c65b9dfa-2e9b-4d30-9b60-b5bbc3021b56

+++ src/scripting.python/PythonMethodAdapter.cpp
@@ -20,7 +20,7 @@ bool parseArguments(PyObject * args,  std::vector<Argument> & arguments, std::ve
   }
   // else use passed types to construct argument
   if(n!=argumentTypes->size())return false;
-  for(int i=0; i < argumentTypes->size(); i++){
+  for(size_t i=0; i < argumentTypes->size(); i++){
     auto argType = argumentTypes->at(i);
     auto argPythonObject = PyTuple_GET_ITEM(args,i);
     auto argument = pythonObjectToArgument(argPythonObject,argType);
